gemini_live_bridge/requirements.txt
websockets>=12.0
numpy>=1.26
scipy>=1.12
aiohttp>=3.9
speexdsp-python>=0.1.0
